% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trawl_acf.R
\name{acf_trawl}
\alias{acf_trawl}
\title{R-only Trawl ACF function approximation.}
\usage{
acf_trawl(
  h,
  alpha,
  beta,
  kappa,
  rho,
  delta = 0.5,
  end_seq = 50,
  type = "exp",
  cov = T
)
}
\arguments{
\item{h}{ACF horizon.}

\item{alpha}{Gamma \code{alpha} (or shape) parameter.}

\item{beta}{Gamma \code{beta} (or rate) parameter.}

\item{kappa}{Extreme frequency control parameter.}

\item{rho}{Trawl parameter(s).}

\item{delta}{Discrete time step.}

\item{end_seq}{Horizon cut-off for integral approximation.}

\item{type}{Trawl type (e.g. \verb{"exp}, "\code{sum_exp}", etc.).}

\item{cov}{Boolean, covariance (if \code{TRUE}) or correlation.
Defaults to covariance.}
}
\value{
Trawl autocorrelation values with R-only functions.
}
\description{
R-only Trawl ACF function approximation.
}
