% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trawl_acf.R
\name{acf_trawl_single}
\alias{acf_trawl_single}
\title{R-only Single Trawl ACF calculations.}
\usage{
acf_trawl_single(
  h,
  alpha,
  beta,
  rho,
  kappa,
  delta = 0.1,
  end_seq = 50,
  type = "exp",
  cov = F
)
}
\arguments{
\item{h}{Single ACF horizon.}

\item{alpha}{Gamma \code{alpha} (or shape) parameter.}

\item{beta}{Gamma \code{beta} (or rate) parameter.}

\item{rho}{Trawl parameter(s).}

\item{kappa}{Extreme frequency control parameter.}

\item{delta}{Discrete time step.}

\item{end_seq}{Horizon cut-off for integral approximation.}

\item{type}{Trawl type (e.g. \verb{"exp}, "\code{sum_exp}", etc.).}

\item{cov}{Boolean, correlation or covariance (if \code{FALSE}).
Defaults to correlation.}
}
\value{
Trawl autocorrelation values with R-only functions.
}
\description{
R-only Single Trawl ACF calculations.
}
