% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hac.R
\name{autocovariance_matrix}
\alias{autocovariance_matrix}
\title{Autocovariance matrix from vector.}
\usage{
autocovariance_matrix(score, k)
}
\arguments{
\item{score}{Vector series.}

\item{k}{Maximum horizon.}
}
\value{
Autocovariance matrix
}
\description{
Autocovariance matrix from vector.
}
\examples{
d <- 5
n <- 1000
vals <- matrix(norm(n * d), ncol = d)
make_hac(vals, 5)
}
