% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{block_index}
\alias{block_index}
\title{Creates the block indices for a collection of trawl slices.}
\usage{
block_index(n_block, n, vanishing_depth)
}
\arguments{
\item{n_block}{Number of blocks}

\item{n}{Number of samples.}

\item{vanishing_depth}{Approximation depth.}
}
\value{
List of block coordinates
}
\description{
Creates the block indices for a collection of trawl slices.
}
