% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_mle.R
\name{composite_marginal_hac}
\alias{composite_marginal_hac}
\title{HAC estimator of autocovariance for the composite marginal.}
\usage{
composite_marginal_hac(data, params, k = 10, max_length = 100, near_pd = F)
}
\arguments{
\item{data}{Data on which to compute the autocovariance.}

\item{params}{Composite likelihood params.}

\item{k}{Maximum lag.}

\item{max_length}{Maximum length of data used to compute the score functions.}
}
\value{
Composite marginal HAC autocovariance.
}
\description{
HAC estimator of autocovariance for the composite marginal.
}
\examples{
xi <- 1.
kappa <- 9.
sigma <- 1.
n <- 100
p_zero <- 1 - 1. / (1. + kappa)
test_column <- 2

set.seed(42)
zeroes <- stats::runif(n = n, min = 0, max = 1)
test_samples <- evir::rgpd(n = n, xi = xi, mu = 0, beta = sigma)
test_samples[which(zeroes < p_zero)] <- 0.0
composite_marginal_hac(
  data = test_samples, params = c(xi, sigma, kappa),
  k = 3, max_length = 50
)
}
