% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trawl_acf.R
\name{cpp_acf_trawl}
\alias{cpp_acf_trawl}
\title{Rcpp-accelerated Trawl ACF function approximation.}
\usage{
cpp_acf_trawl(
  h,
  alpha,
  beta,
  kappa,
  rho,
  delta = 0.5,
  type = "exp",
  end_seq = 50,
  cov = T
)
}
\arguments{
\item{h}{ACF horizon.}

\item{alpha}{Gamma \code{alpha} (or shape) parameter.}

\item{beta}{Gamma \code{beta} (or rate) parameter.}

\item{kappa}{Extreme frequency control parameter.}

\item{rho}{Trawl parameter(s).}

\item{delta}{Discrete time step.}

\item{type}{Trawl type (e.g. \code{"exp"}, \code{"sum_exp"}, etc.).}

\item{end_seq}{Horizon cut-off for integral approximation.}

\item{cov}{Boolean, correlation or covariance (if \code{FALSE}).
Defaults to correlation.}
}
\value{
Trawl autocorrelation values with Cpp-accelerated functions.
}
\description{
Rcpp-accelerated Trawl ACF function approximation.
}
