% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_case_one_one}
\alias{cpp_case_one_one}
\title{Cpp pairwise likelihood 1-1 (one-one).}
\usage{
cpp_case_one_one(xs, alpha, beta, kappa, b_1, b_2, b_3)
}
\arguments{
\item{xs}{Data.}

\item{alpha}{Gamma shape.}

\item{beta}{Gamma rate.}

\item{kappa}{Exceedance controlling parameter.}

\item{b_1}{B one.}

\item{b_2}{B two.}

\item{b_3}{B three.}
}
\description{
Cpp pairwise likelihood 1-1 (one-one).
}
