% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_case_separator}
\alias{cpp_case_separator}
\title{Cpp pairwise likelihood separator that chooses which routine to use.}
\usage{
cpp_case_separator(xs, alpha, beta, kappa, b_1, b_2, b_3)
}
\arguments{
\item{xs}{Data.}

\item{alpha}{Gamma shape.}

\item{beta}{Gamma rate.}

\item{kappa}{Exceedance controlling parameter.}

\item{b_1}{B one.}

\item{b_2}{B two.}

\item{b_3}{B three.}
}
\value{
PL likelhood values for \code{xs}.
}
\description{
Cpp pairwise likelihood separator that chooses which routine to use.
}
