% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_case_zero_zero}
\alias{cpp_case_zero_zero}
\title{Cpp pairwise likelihood 0-0 (zero-zero).}
\usage{
cpp_case_zero_zero(alpha, beta, kappa, b_1, b_2, b_3)
}
\arguments{
\item{alpha}{Gamma shape.}

\item{beta}{Gamma rate.}

\item{kappa}{Exceedance controlling parameter.}

\item{b_1}{B one.}

\item{b_2}{B two.}

\item{b_3}{B three.}
}
\value{
pairwise likelihood 0-0 (zero-zero).
}
\description{
Cpp pairwise likelihood 0-0 (zero-zero).
}
