% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_mle.R
\name{custom_likelihood}
\alias{custom_likelihood}
\title{Likelihood function of parameters for fixed data.}
\usage{
custom_likelihood(data)
}
\arguments{
\item{data}{Data on which to compute likelihood.}
}
\value{
Custom likelihood function of parameters.
}
\description{
Likelihood function of parameters for fixed data.
}
\examples{
set.seed(42)
xi <- 1.
kappa <- 9.
sigma <- 1.
n <- 1000
p_zero <- 1 - 1 / (1 + kappa)
zeroes <- stats::runif(n = n, min = 0, max = 1)
test_samples <- evir::rgpd(n = n, xi = xi, mu = 0, beta = sigma)
test_samples[which(zeroes < p_zero)] <- 0.0
cm_mle <- custom_likelihood(data = test_samples)
}
