% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{exceedances_simulation}
\alias{exceedances_simulation}
\title{Simulates a path from the Latent Trawl Model for Extremes (LTME) process
(Noven et al. (2018) \doi{10.21314/JEM.2018.179})}
\usage{
exceedances_simulation(
  params,
  n,
  vanishing_depth,
  type,
  m = 1,
  parametrisation = "standard",
  parallel = F,
  algo = "standard"
)
}
\arguments{
\item{params}{Trawl parameter(s).}

\item{n}{Number of samples.}

\item{vanishing_depth}{Approximation depth.}

\item{type}{Trawl type (e.g. \code{"exp"}, \code{"sum_exp"}).}

\item{m}{number of paths to use: if \code{algo = "standard"} or number of
supporting paths if \code{algo = "cross"}.}

\item{parametrisation}{Parametrisation to use, defaults is standard
\code{GPD} parameters \verb{(xi, sigma)}.}

\item{parallel}{Whether to parallelise, defaults to \code{FALSE}. NOTE: only works
with \code{"standard"} algorithm.}

\item{algo}{Simulation algorithm to use (e.g \code{"standard"}, \code{"cross"},
\code{"corr_unif"}, \code{"dynamic_latent"}, \code{"dynamic_uniform"}).}
}
\value{
Simulate path from trawl process.
}
\description{
Simulates a path from the Latent Trawl Model for Extremes (LTME) process
(Noven et al. (2018) \doi{10.21314/JEM.2018.179})
}
\examples{
alpha <- 1
beta <-1
n <- 100
vanishing_depth <- 30
trawl_parameter <- 0.2
type <- "exp"
exceedances_simulation(trawl_parameter, n, vanishing_depth, type)
}
