% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Fitting the LTME or EV Trawl model.}
\usage{
fit(data, depth, method, mode, type, bounds, parallel = F)
}
\arguments{
\item{data}{Data to fit.}

\item{depth}{Depth to fit.}

\item{method}{Selecting \code{"GMM"} or \code{"PL"}.}

\item{mode}{Either one-stage/full (\code{"full"}) or two-stage (\code{"two-stage"}).}

\item{type}{Trawl type (e.g. \code{"exp"}, \code{"sum_exp"}, etc.)}

\item{bounds}{Parameters bounds. \code{"config"} uses trawl \code{cfg}, \code{"multiplier"}
uses scaled initial guess.}

\item{parallel}{Boolean, defaults to \code{FALSE}.}
}
\value{
List of fitted parameters, hyperparameters, etc.
}
\description{
Fitting the LTME or EV Trawl model.
}
\examples{
n <- 2000
test_column <- 2
data <- pollution_data[seq_len(n), test_column]
depth <- 5
ev_trawl_fit(data, depth, method = "GMM")
}
