% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{gamma_grid}
\alias{gamma_grid}
\title{Grid of Gamma-distributed trawl slices.}
\usage{
gamma_grid(alpha, beta, n, vanishing_depth, trawl_parameter, type = "exp")
}
\arguments{
\item{alpha}{Gamma shape parameter.}

\item{beta}{Gamma rate parameter.}

\item{n}{Number of samples.}

\item{vanishing_depth}{Approximation depth.}

\item{trawl_parameter}{Trawl parameter(s).}

\item{type}{Trawl type (e.g. \code{"exp"}, \code{"sum_exp"}).}
}
\value{
Gamma-distribute trawl slices.
}
\description{
Grid of Gamma-distributed trawl slices.
}
