% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{gamma_orchestra}
\alias{gamma_orchestra}
\title{Orchestrate the Gamma grid.}
\usage{
gamma_orchestra(scaled_gamma_grid, parallel = T)
}
\arguments{
\item{scaled_gamma_grid}{Gamma grid.}

\item{parallel}{Whether to parallelise, defaults to \code{TRUE}.}
}
\value{
Vector of Gamma-distributed grid.
}
\description{
Orchestrate the Gamma grid.
}
