% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate_shapes}
\alias{generate_shapes}
\title{Generate shape parameters for Gamma-distributed trawl slices.}
\usage{
generate_shapes(alpha, n, vanishing_depth, trawl_parameter, type = "exp")
}
\arguments{
\item{alpha}{Gamma shape parameter.}

\item{n}{Number of samples.}

\item{vanishing_depth}{Approximation depth.}

\item{trawl_parameter}{Trawl parameter(s).}

\item{type}{Trawl type (e.g. \code{"exp"}, \code{"sum_exp"}).}
}
\value{
Gamma-distribute trawl slices.
}
\description{
Generate shape parameters for Gamma-distributed trawl slices.
}
