% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trawl_utils.R
\name{get_trawl_env}
\alias{get_trawl_env}
\title{Returns a trawl environment with numerical utils functions that implements
\code{trawl_b_one}, \code{trawl_b_two} and \code{trawl_b_three}.}
\usage{
get_trawl_env(type)
}
\arguments{
\item{type}{Trawl type (e.g. \code{"exp"}, \code{"sum_exp"}, \code{"sup_ig"}, \code{"gamma"}).}
}
\value{
An R environment.
}
\description{
Returns a trawl environment with numerical utils functions that implements
\code{trawl_b_one}, \code{trawl_b_two} and \code{trawl_b_three}.
}
\examples{
get_trawl_env("exp")
}
