% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trawl_utils.R
\name{get_trawl_envs_list}
\alias{get_trawl_envs_list}
\title{Lists of available trawl functions.}
\usage{
get_trawl_envs_list()
}
\value{
Vector of trawl names.
}
\description{
Lists of available trawl functions.
}
\examples{
get_trawl_envs_list()
}
