% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trawl_utils.R
\name{get_trawl_functions}
\alias{get_trawl_functions}
\title{Obtaining the three functions necessary for the trawls.}
\usage{
get_trawl_functions(type)
}
\arguments{
\item{type}{Trawl type (e.g. \code{"exp"}, \code{"sum_exponential_trawl"}, etc).}
}
\value{
A vector of three functions to define trawl intersections,
which depend on the trawl environment.
}
\description{
Obtaining the three functions necessary for the trawls.
}
\examples{
get_trawl_functions("exp") # Exponential trawl
}
