% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trawl_utils.R
\name{get_trawl_params_config}
\alias{get_trawl_params_config}
\title{Returns triplet list (n_params, lower, upper) for each trawl type.}
\usage{
get_trawl_params_config(type)
}
\arguments{
\item{type}{Trawl type (e.g. \code{"exp"}, \code{"sum_exponential_trawl"}, etc).}
}
\value{
A list of three vectors \code{list(n_params, lower, upper)} to define
trawl intersections parameters.
}
\description{
Returns triplet list (n_params, lower, upper) for each trawl type.
}
\examples{
get_trawl_params_config("exp") # Exponential trawl cfg
}
