% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{grid_foundations}
\alias{grid_foundations}
\title{Returns a (sparse) triangular matrix and plug values in from top to bottom,
left to right.}
\usage{
grid_foundations(n, vanishing_depth, values = 1)
}
\arguments{
\item{n}{Number of time steps.}

\item{vanishing_depth}{Appeoximation depth.}

\item{values}{Either scalar or vector to input in matrix.}
}
\value{
Sparse matrix with \code{values} included.
}
\description{
Returns a (sparse) triangular matrix and plug values in from top to bottom,
left to right.
}
\examples{
n <- 100
vd <- 50
grid_foundations(n, vd)
}
