% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.R
\name{pl_ck_positive}
\alias{pl_ck_positive}
\title{Checking that vectors is positive element-wise.}
\usage{
pl_ck_positive(elems)
}
\arguments{
\item{elems}{Elements.}
}
\description{
Checking that vectors is positive element-wise.
}
