% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.R
\name{pl_stand_trawl_terms}
\alias{pl_stand_trawl_terms}
\title{Standardised trawl's shape or alpha parameter.}
\usage{
pl_stand_trawl_terms(alpha, elems)
}
\arguments{
\item{alpha}{Alpha parameter.}

\item{elems}{Elements.}
}
\description{
Standardised trawl's shape or alpha parameter.
}
