% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translator.R
\name{parametrisation_translator}
\alias{parametrisation_translator}
\title{Transforms one parametrisation into another one.
From Gamma-inspired parameters (\code{noven}) to standard (i.e. \verb{(xi, sigma)}).}
\usage{
parametrisation_translator(params, parametrisation, target = "noven")
}
\arguments{
\item{params}{Parameters to transform.}

\item{parametrisation}{Given parametrisation.}

\item{target}{Target parametrisation (in \code{"standard"}, \code{"noven"} or
\code{transform}).}
}
\value{
Transformed parameters.
}
\description{
Transforms one parametrisation into another one.
From Gamma-inspired parameters (\code{noven}) to standard (i.e. \verb{(xi, sigma)}).
}
\examples{
parametrisation_translator(c(1,1), "noven", "standard")
}
