% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollution_data.R
\docType{data}
\name{pollution_data}
\alias{pollution_data}
\title{Air Pollutant Concentration on Marylebone Road, London, UK.}
\format{
A data frame with 157,710 rows and 6 variables:
\describe{
\item{O3}{Ozone air concentration, in µg/m3}
\item{CO}{Carbon Monoxide air concentration, in µg/m3}
\item{NO}{Nitric Oxide air concentration, in µg/m3}
\item{NO2}{Nitric Dioxide air concentration, in µg/m3}
\item{PM10}{Particulate Matter below 10mm concentration, in µg/m3}
\item{SO2}{Sulphur Dioxide air concentration, in µg/m3}
}
}
\source{
\url{
   https://uk-air.defra.gov.uk/data/site-data?f_site_id=MY1&view=last_hour
}
}
\usage{
pollution_data
}
\description{
A dataset containing the air pollutant concentration of six molecules:
Ozone (O3), Carbon Monoxide (CO), Nitric Oxide (NO), Nitric Dioxide (NO2),
Particulate Matter below 10mm (PM10), Sulphur Dioxide (SO2).
}
\keyword{datasets}
