% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{print_vanishing_coverage}
\alias{print_vanishing_coverage}
\title{Returns the coverage of the trawl set approximation.}
\usage{
print_vanishing_coverage(
  trawl_parameter,
  vanishing_depth,
  type = "exp",
  get_value = F
)
}
\arguments{
\item{trawl_parameter}{Trawl parameter(s).}

\item{vanishing_depth}{Approximation depth.}

\item{type}{Trawl type (e.g. \code{"exp"}, \code{"sum_exp"}).}

\item{get_value}{Prints or returns the value. Defaults to \code{FALSE}.}
}
\value{
Trawl set approximation (print or value).
}
\description{
Returns the coverage of the trawl set approximation.
}
\examples{
print_vanishing_coverage(trawl_parameter = 0.3, vanishing_depth = 10)
}
