% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{square_moment}
\alias{square_moment}
\title{Computes square moment in trawl gamma-exponential mixture}
\usage{
square_moment(xs, delta, beta, b_oh, b_o_exc_h)
}
\arguments{
\item{xs}{Grid to integrate on.}

\item{delta}{Grid mesh size}

\item{beta}{GPD scale parameter in the sense of Noven et al.}

\item{b_oh}{\itemize{
\item alpha * mu(A inter A_h) / mu  (A)
}}

\item{b_o_exc_h}{\itemize{
\item alpha * mu(A excl. A_h) / mu  (A)
}}
}
\value{
Square moment value
}
\description{
Computes square moment in trawl gamma-exponential mixture
}
