% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{sub_sample_fit}
\alias{sub_sample_fit}
\title{Fitting the LTME or EV Trawl model, potentially on block subsamples.}
\usage{
sub_sample_fit(
  data,
  sample_length,
  depth,
  method,
  mode,
  type,
  bounds,
  trials,
  parallel = F,
  seed = 42
)
}
\arguments{
\item{data}{Data to fit.}

\item{sample_length}{Length of block subsamples.}

\item{depth}{Depth to fit.}

\item{method}{Selecting \code{"GMM"} or \code{"PL"}.}

\item{mode}{Either one-stage/full (\code{"full"}) or two-stage (\code{"two-stage"}).}

\item{type}{Trawl type (e.g. \code{"exp"}, \code{"sum_exp"}, etc.)}

\item{bounds}{Parameters bounds. \code{"config"} uses trawl cfg, \code{"multiplier"}
uses scaled initial guess.}

\item{cl}{Parallel cluster, defaults to \code{NULL}.}
}
\value{
Vector of fitted parameters.
}
\description{
Fitting the LTME or EV Trawl model, potentially on block subsamples.
}
\examples{
n <- 2000
test_column <- 2
data <- pollution_data[seq_len(n), test_column]
depth <- 5
ev_trawl_fit(data, depth, method = "GMM")
}
