% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_transformation.R
\name{transformation_jacobian}
\alias{transformation_jacobian}
\title{Jacobian of the transformation function.}
\usage{
transformation_jacobian(params_std)
}
\arguments{
\item{params_std}{Standard parametrisation parameters \verb{(xi, sigma)}.}
}
\value{
Jacobian of the transformation function.
}
\description{
Jacobian of the transformation function.
}
\examples{
xi <- .1
sig <- .5
kap <- 19.
transformation_jacobian(c(xi, sig, kap))
}
