% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_transformation.R
\name{transformation_map}
\alias{transformation_map}
\title{Transforms data from \code{GPD(1, 1 + kappa)} to \code{GPD(xi, sigma)}.}
\usage{
transformation_map(x, params_std)
}
\arguments{
\item{x}{Data to transform.}

\item{params_std}{Vector \code{c(xi, sigma, kappa)}.}
}
\value{
Vector of \code{GPD(xi, sigma)} data.
}
\description{
Transforms data from \code{GPD(1, 1 + kappa)} to \code{GPD(xi, sigma)}.
}
\examples{
xi <- 1
sig <- 3
kap <- 2
tmp <- evir::rgpd(100, xi = 1, beta = 1 + kap)
transformation_map(tmp, c(xi, sigma))
}
