% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_transformation.R
\name{transformation_map_inverse}
\alias{transformation_map_inverse}
\title{Inverse transformation function.}
\usage{
transformation_map_inverse(x, params_std)
}
\arguments{
\item{x}{Data to transform}

\item{params_std}{Standard parametrisation parameters \verb{(xi, sigma)}.}
}
\value{
\code{GPD(1, 1 + kappa)} data.
}
\description{
Inverse transformation function.
}
\examples{
n <- 10
xi <- .1
sig <- .5
kap <- 19.
data <- evir::rgpd(n, xi = xi, beta = sig)
transformation_map_inverse(data, c(xi, sig, kap))
}
