% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{trawl_simulation}
\alias{trawl_simulation}
\title{Simulates a path from a Trawl process.}
\usage{
trawl_simulation(
  alpha,
  beta,
  n,
  vanishing_depth,
  trawl_parameter,
  type,
  parallel = F
)
}
\arguments{
\item{alpha}{Gamma shape parameter.}

\item{beta}{Gamma rate parameter.}

\item{n}{Number of samples.}

\item{vanishing_depth}{Approximation depth.}

\item{trawl_parameter}{Trawl parameter(s).}

\item{type}{Trawl type (e.g. \code{"exp"}, \code{"sum_exp"}).}

\item{parallel}{Whether to parallelise, defaults to \code{FALSE}.}
}
\value{
Simulate path from trawl process.
}
\description{
Simulates a path from a Trawl process.
}
\examples{
alpha <- 1
beta <-1
n <- 100
vanishing_depth <- 30
trawl_parameter <- 0.2
type <- "exp"
trawl_simulation(alpha, beta, n, vanishing_depth, trawl_parameter, type)
}
