% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hac.R
\name{make_hac}
\alias{make_hac}
\title{Autocorrelation-proof estimates using HAC estimator.}
\usage{
make_hac(acf_matrices, near_pd = F)
}
\arguments{
\item{acf_matrices}{Matrices to use in the estimators.}

\item{near_pd}{Boolean, nearest positive definite matrix from HAC estimate.}
}
\value{
HAC-transformed matrix
}
\description{
Autocorrelation-proof estimates using HAC estimator.
}
\examples{
d <- 5
mats <- array(0, c(d, d, d))
mats[1,,] <- diag(d)
make_hac(mats)
}
